Added Interface for Transaction Service
